Add permissions for list view
Refactor the messageListServer to take a views.Client instead of a
twilio.Client. Add support for the GetNextPage and Messages.GetPage calls in
the views client. Update the template so we check whether we can access fields
before displaying them.

Also moves the footer duration/render information to the base.html template,
and make it stick to the bottom of the page if we have fewer resources to
display than normal.