Add code to hide images by default / show them on click
This adds the site's first snippet of Javascript. I'm not happy about it but
the CSS-based alternative involves hiding a checkbox and then transitioning the
div *next* to it, which is also pretty bad. Since we have it I got a little
fancy and added some nice transitions for images.

Adds some basic commands to the binary - it understands "help", "version", and
"serve" now, and has a nicer help message.