Show error code / error message
If a message has an error code/error message, show them on the instance page.
Also add links from the Status page of the list view. I'm not super happy with
the list view - I think we can find a way to show more information about the
error on that page. I'll keep thinking about it.