Implement Call list view
Adds a bunch of scaffolding for viewing/displaying Calls, in views and in the
server. Splits out some code that was messages-specific into page.go files. The
Calls styling is a little messier than the Messages styling because you can
call SIP numbers and Client numbers. I don't have a good solution for this but
imagine we'll have to truncate these fields with an ellipsis in the list view
or enforce wrapping with a max-width at some point.