Implement Google auth
Add configuration so you can specify a "google" auth scheme, and redirect
people to Google/set a cookie based on their correctly identifying with Google
login.

We still need to implement multi-user and a logout function. Also right now
anyone who can log in via Google can log in to the website, we need to
implement an "approved users" list.

We also probably need a way to render a 400 Bad Request if something goes wrong
in the authentication flow.