Cache list results
If a user is browsing through pages, fetch the next page and cache it before
they get to it. This way we can serve from the cache (fast) instead of waiting
to retrieve the results from the Twilio API (slow) when the user requests the
next page.

There's a race between page 1 fetching page 2 and putting it in the cache, and
the user requesting page 2. We avoid this by using singleflight to only make
one request at a time for page 2, and checking the cache inside singleflight
before hitting the Twilio API.

The cache won't be as effective if you are running multiple instances of
Logrole since incoming requests will be split across the N machines. We can set
up groupcache to run across all machines but that'll take more work.