Implement click-to-copy
Every phone number now has a little clipboard next to it with a click-to-copy
clipboard. Pulls out the number display into a snippet so we can reuse it in
every template.

Strips the "+1 " prefix from US numbers - international numbers will still
have the prefix, but US numbers don't need it. Eventually I can make this
configurable.