Color Twilio phone numbers in green
This required a few different pieces to work:

- We have to fetch phone numbers from the Twilio API and store them on the
views client. Not too bad though we should add an API for this besides in
views.NewClient.

- We had to put templates on each Server object instead of making them global.
I think this will be beneficial in the long run. Though we should switch each
asset back to being loaded in init & being a package-level variable, this
way we can share them across all templates that need them and avoid extra
allocations.

- CSS for styling.

- A new IsTwilioNumber function which we pass to each template that displays
  phone numbers.

Updates twilio-go to 0.29 which contains support for PhoneNumbers and a better
readme.