Skeleton of a conference list page
Fix an error in template rendering where we were reusing a bytes.Buffer
without resetting it first, which was a mistake that led to partially-rendered
templates. Not good!

Adds a server endpoint that fetches the first page of conference results, and
then displays "Hello World". Next up rendering that conference data on the
page.

Updates twilio-go to get the new Conferences resource.