Load timezones in menu bar from configuration
Allow users to specify a list of timezones, e.g. "Asia/Singapore,Asia/Tokyo",
and display the users timezones in the menu bar. Defaults to the four US
timezones.

Move the Settings object to the config directory, it makes more sense there.
The only bad thing is the Authenticator definition also has to move there, but
we need to keep the GoogleAuthenticator in the services package since it
renders stuff. Maybe we can refactor the definition here or use another
interface to work around the import cycle problems.