Add benchmarks
Update handlers to 0.26 so we can have more control over the HTTP request
logger. Specifically we want to disable output when running benchmarks (it
would be useful to benchmark but prints a ton of lines to stdout).

Add two benchmarks - one for rendering the login page and one for rendering the
message list page against a dummy Twilio server. They are pretty slow but it's
good to have them. Run the benchmarks in Travis CI.

Cache no longer initializes its own logger or has a debug variable, it pulls
both of these from the system-wide settings.

Moves the Harness to its own class but given its dependencies it may be better
suited just living in the server package.