Implement caching for all resources
Show whether a result was returned from the cache, and if so how old that
result is. Alters services.Duration to show fewer bits after the decimal if the
Duration is larger than one second. Updates the UI to show whether a result was
served from the cache or not.