Add call/message summaries on PN page
When you fetch a phone number, get all of the calls/messages to and from that
number, and display them on the page.

This barely works in the wide view, I hope it's alright. I had to shorten the
date view to make it fit, if your message is over a year old I think it's okay
to truncate the exact time it was sent. I can change this back if need be.