Fix list filters
Now they all fit on one level on a 13" Macbook, with the exception of Alerts,
which I put on two lines. The styling for Alerts could be better than it is.

Adds a PhoneNumber filter on the numbers list view, so you can filter for
a number (or part of one).