Update kevinburke/rest to latest master
At some point the signature of RegisterErrorHandler changed from (int,
HandlerFunc) to (int, Handler), which is more flexible. Unfortunately the code
here didn't change.

We didn't notice the build failure because an older version of kevinburke/rest
was being vendored. In any case, let's fix it and update the vendored
dependency.

Fixes #21.